passage de l'ensemble des API en Long plutôt que Integer pour la clé des entités
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@1057 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core/owsi-core-components/owsi-core-component-jpa-more/src/main/java/fr/openwide/core/jpa/more/business/audit/dao/IAuditDao.java
@@ -25,7 +25,7 @@ import fr.openwide.core.jpa.more.business.audit.model.AbstractAudit;
 /**
  * <p>DAO concernant les {@link AbstractAudit}.</p>
  */
-public interface IAuditDao<T extends AbstractAudit> extends IGenericEntityDao<Integer, T> {
+public interface IAuditDao<T extends AbstractAudit> extends IGenericEntityDao<Long, T> {
 
 	/**
 	 * Renvoie l'entité correspondant à la classe et l'identifiant demandés.
@@ -35,7 +35,7 @@ public interface IAuditDao<T extends AbstractAudit> extends IGenericEntityDao<In
 	 * @return entité
 	 */
 	GenericEntity<?, ?> getGenericEntity(Class<? extends GenericEntity<?, ?>> clazz,
-			Integer id);
+			Long id);
 
 	/**
 	 * Renvoie la liste des Audits émis par cet utilisateur.
