+++ owsi-core/owsi-core-components/owsi-core-component-jpa-security/src/main/java/fr/openwide/core/jpa/security/business/authority/dao/AuthorityDaoImpl.java
@@ -6,7 +6,7 @@ import fr.openwide.core.jpa.business.generic.dao.GenericEntityDaoImpl;
 import fr.openwide.core.jpa.security.business.authority.model.Authority;
 
 @Repository("authorityDao")
-public class AuthorityDaoImpl extends GenericEntityDaoImpl<Integer, Authority> implements IAuthorityDao {
+public class AuthorityDaoImpl extends GenericEntityDaoImpl<Long, Authority> implements IAuthorityDao {
 	
 	public AuthorityDaoImpl() {
 		super();
