+++ owsi-core/owsi-core-components/owsi-core-component-jpa-security/src/main/java/fr/openwide/core/jpa/security/business/person/dao/IPersonGroupDao.java
@@ -5,6 +5,6 @@ import fr.openwide.core.jpa.security.business.person.model.AbstractPerson;
 import fr.openwide.core.jpa.security.business.person.model.AbstractPersonGroup;
 
 public interface IPersonGroupDao<G extends AbstractPersonGroup<G, P>, P extends AbstractPerson<P>>
-		extends IGenericEntityDao<Integer, G> {
+		extends IGenericEntityDao<Long, G> {
 	
 }
