+++ owsi-core/owsi-core-components/owsi-core-component-wicket-more/src/main/java/fr/openwide/core/wicket/more/markup/html/list/GenericPortfolioPanel.java
@@ -13,7 +13,7 @@ import fr.openwide.core.wicket.more.markup.html.template.js.jquery.plugins.tipsy
 import fr.openwide.core.wicket.more.markup.html.template.js.jquery.plugins.tipsy.TipsyBehavior;
 
 
-public abstract class GenericPortfolioPanel<E extends GenericEntity<Integer, ?>>
+public abstract class GenericPortfolioPanel<E extends GenericEntity<Long, ?>>
 		extends AbstractGenericItemListPanel<E> {
 
 	private static final long serialVersionUID = 1L;
