+++ owsi-core/owsi-core-examples/wicket-showcase/wicket-showcase-core/src/main/java/fr/openwide/core/showcase/core/security/acl/service/ShowcaseAclServiceImpl.java
@@ -23,7 +23,7 @@ public class ShowcaseAclServiceImpl extends AbstractCoreAclServiceImpl {
 
 	@Override
 	protected List<AccessControlEntry> getAccessControlEntriesForEntity(
-			CoreAcl acl, GenericEntity<Integer, ?> objectIdentityEntity)
+			CoreAcl acl, GenericEntity<Long, ?> objectIdentityEntity)
 			throws NotFoundException {
 		return new ArrayList<AccessControlEntry>();
 	}
